+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.1.4-alpha] - 2026-06-
+## [0.1.5-alpha] - 2026-
+### Added
+- Player can now choose his deck
+- save player data and remove then in settings window
+- number of card remaining in the deck
+- add color for the next level
+- if the player hasn't selected a deck when starting a level, a window will force the player to choose one.
+
+### Fixed
+- Each click on the deck remove a card even if the handslot is full
+
+
+## [0.1.4-alpha] - 2026-06-07
 ### Added
 - story system with json file
 - log system with UI button